Move to JSON-RPC based transport in preparation of TCP support (#2)
* Update copyright

* Make registerProcess automagic (backwards compatible)

Make the use of `.registerProcess()` obsolete.
`registerProcess` will emit a warning but do nothing further.
Process regstration is done in the constructor.

* Rename ClientRPC -> WorkerRPC (backwards compatible)

Renamed `ClientRPC` to `WorkerRPC` to reenforce the meaning, that it is intended for worker to master communication.

* Rename callOnClients -> callOnWorkers

Rename `callOnClient` to `callOnWorkers` to make naming resemble the actual communication path to all attached workers, where as clients where pretty ambiguous.

* make `installHandlerOn` only accept `ChildProcess`es

Installing a handler on `Worker` is redundant and caused all sorts of weiredness. This should be resolved by this.
It will not *prevent* anybody from handing over a `Worker` Object, but it will show up in the TSC and make the user aware of this.

* Make registerProcess automagic in ChildRPC, too (backwards compatible)

Make the use of `.registerProcess()` obsolete.
`registerProcess` will emit a warning but do nothing further.
Process regstration is done in the constructor of the parent.

* Switch to JSON-RPC-like call structure.

Not completely to spec, yet. As error codes are missing and lots of error handling in general has to be improved.

* Bump version to 0.9.0

Reasoning:
Close to being 1.0. Before the switch to JSON-RPC this library was used in production workloads, and regression-testing showed, that this library still behaves the same (apart from the fixed double response issue).

* Add a demo

* add `callOnMaster<T>` to the `RPC` base and `MasterRPC` for user comfort